requirements.txt
# Nikki Flower & Cake Requirements

# Core Django Framework
Django==6.0.2

# Database
psycopg2-binary==2.9.9  # Required for PostgreSQL in production
Pillow==12.1.1  # Required for image handling

# Environment Variables
python-decouple==3.8

# Caching
django-redis==6.0.0
redis==7.2.1

# Static Files
whitenoise==6.12.0

# Debugging
django-debug-toolbar==6.2.0

# Payment Processing
razorpay==1.4.0

# Async Tasks (Future Implementation)
celery==5.4.0
redis==7.2.1  # For Celery broker

# API Framework (Future Implementation)
djangorestframework==3.15.2

# CORS (For future mobile app/API)
django-cors-headers==4.3.1

# Email
django-anymail==10.3  # For transactional emails

# File Storage (Future cloud storage)
django-storages==1.14.2
boto3==1.34.150  # For AWS S3

# Security
django-csp==3.8  # Content Security Policy